requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
cloudinary==1.31.0
Django==4.1.5
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-filter==22.1
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
Markdown==3.4.1
Pillow==9.4.0
pylibmagic==0.2.2
python-decouple==3.7
python-magic==0.4.27
pytz==2022.7.1
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
whitenoise==6.3.0
